+++ src/main/java/emu/grasscutter/game/avatar/Avatar.java
@@ -1,15 +1,14 @@
 package emu.grasscutter.game.avatar;
 
-import static emu.grasscutter.config.Configuration.GAME_OPTIONS;
-
 import dev.morphia.annotations.*;
 import emu.grasscutter.GameConstants;
 import emu.grasscutter.data.GameData;
 import emu.grasscutter.data.binout.OpenConfigEntry;
 import emu.grasscutter.data.binout.OpenConfigEntry.SkillPointModifier;
 import emu.grasscutter.data.common.FightPropData;
-import emu.grasscutter.data.excels.*;
+import emu.grasscutter.data.excels.EquipAffixData;
 import emu.grasscutter.data.excels.ItemData.WeaponProperty;
+import emu.grasscutter.data.excels.ProudSkillData;
 import emu.grasscutter.data.excels.avatar.AvatarData;
 import emu.grasscutter.data.excels.avatar.AvatarSkillData;
 import emu.grasscutter.data.excels.avatar.AvatarSkillDepotData;
@@ -41,15 +40,18 @@ import emu.grasscutter.net.proto.TrialAvatarInfoOuterClass.TrialAvatarInfo;
 import emu.grasscutter.server.packet.send.*;
 import emu.grasscutter.utils.ProtoHelper;
 import it.unimi.dsi.fastutil.ints.*;
-import java.util.*;
-import java.util.stream.Stream;
-import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.val;
 import org.bson.types.ObjectId;
+
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import java.util.*;
+import java.util.stream.Stream;
+
+import static emu.grasscutter.config.Configuration.GAME_OPTIONS;
 
 @Entity(value = "avatars", useDiscriminator = false)
 public class Avatar {
@@ -316,7 +318,7 @@ public class Avatar {
         var candSkillDepotId = candSkillDepotIdsList.get(candSkillDepotIndex);
 
         // Sanity checks for skill depots
-        val skillDepot = GameData.getAvatarSkillDepotDataMap().get(candSkillDepotId);
+        val skillDepot = GameData.getAvatarSkillDepotDataMap().get((int) candSkillDepotId);
         if (skillDepot == null || skillDepot.getId() == skillDepotId) {
             return false;
         }
